Artifacts from https://github.com/floodlight/loxigen/commit/b675351e7c349198eef0c5298906b3d3919adf0e (Branch master)
Loxigen Head commit floodlight/loxigen@b675351e7c349198eef0c5298906b3d3919adf0e
commit b675351e7c349198eef0c5298906b3d3919adf0e
Merge: 1393a6a 4a41a4f
Author: abat <abat@bigswitch.com>
Date:   Fri Mar 9 23:30:28 2018 +0000

    Merge into master from pull request #587:
    [SWL-4374] Changed the FEC config enum type to uint32_t (https://github.com/floodlight/loxigen/pull/587)

+++ loci/inc/loci/loci_classes.h
@@ -28576,10 +28576,10 @@ extern void of_port_desc_prop_bsn_forward_error_correction_exp_type_get(
 
 extern void of_port_desc_prop_bsn_forward_error_correction_configured_set(
     of_port_desc_prop_bsn_forward_error_correction_t *obj,
-    uint16_t configured);
+    uint32_t configured);
 extern void of_port_desc_prop_bsn_forward_error_correction_configured_get(
     of_port_desc_prop_bsn_forward_error_correction_t *obj,
-    uint16_t *configured);
+    uint32_t *configured);
 
 extern void of_port_desc_prop_bsn_forward_error_correction_enabled_set(
     of_port_desc_prop_bsn_forward_error_correction_t *obj,
